fix issues after comments
